+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_training/green_screen.dart';
+import 'package:flutter_training/screens/green_screen.dart';
 
 void main() {
   runApp(
