+++ lib/models/weather_kind.dart
