+++ lib/models/weather_request.dart
