+++ lib/models/weather_response.freezed.dart
